3D Surface: Version 6a, Baby!  Yes!
Say hello to my little friend, Version 6a!
New feature:  Face avoidance!  That is right.  Add faces to the end of the Base Geometry list, then set the new property, `AvoidLastXFaces` to tell FreeCAD and 3D Surface to avoid the last X faces in the Base Geometry list!  All other remaining above(at the top) will be CUT, or the rest of the entire model if you like!