Path: Fix edge identification method
If edges are not perfectly identical, and orientation not the same, then existing method failed to identify matching edges correctly.
This new method does a better job of identifying matching edge pairs - to then be removed because they are internal to the overall region to be cleared.
Path: Isolate simple edge-selection algorithm

This simple algorithm removes all shared edges from all faces passed to it and only returns a list of unique edges, not simplified faces.
The adaptive algorithm only accepts a list containing lists of the discretized edge segments for each edge.
Path: Correct indentation.